@@ -1,11 +1,7 @@
 module github.com/yuin/gopher-lua
 
-go 1.17
+go 1.22
 
-require github.com/chzyer/readline v0.0.0-20180603132655-2972be24d48e
+require github.com/chzyer/readline v1.5.1
 
-require (
-	github.com/chzyer/logex v1.1.10 // indirect
-	github.com/chzyer/test v0.0.0-20180213035817-a1ea475d72b1 // indirect
-	golang.org/x/sys v0.0.0-20190204203706-41f3e6584952 // indirect
-)
+require golang.org/x/sys v0.24.0 // indirect
