@@ -37,7 +37,8 @@ func loFindFile(L *LState, name, pname string) (string, string) {
 	messages := []string{}
 	for _, pattern := range strings.Split(string(path), ";") {
 		luapath := strings.Replace(pattern, "?", name, -1)
-		if _, err := L.Stat(luapath); err == nil {
+		if f, err := L.openFile(luapath); err == nil {
+			f.Close()
 			return luapath, ""
 		} else {
 			messages = append(messages, err.Error())
