@@ -8,6 +8,7 @@ import (
 	"context"
 	"fmt"
 	"io"
+	"io/fs"
 	"math"
 	"os"
 	"runtime"
@@ -109,8 +110,8 @@ type Options struct {
 	// If `MinimizeStackMemory` is set, the call stack will be automatically grown or shrank up to a limit of
 	// `CallStackSize` in order to minimize memory usage. This does incur a slight performance penalty.
 	MinimizeStackMemory bool
-	// Load lua files from LuaFileSystem instead of OS file-system.
-	LuaFileSystem LuaFileSystem
+	// FileSystem.
+	LuaFileSystem fs.FS
 }
 
 /* }}} */
@@ -636,28 +637,6 @@ func (rg *registry) IsFull() bool {
 
 /* }}} */
 
-/* luaFileSystem {{{ */
-type LuaFileSystem interface {
-	Open(path string) (io.ReadCloser, error)
-	Stat(luapath string) (os.FileInfo, error)
-}
-
-func (ls *LState) Open(path string) (io.ReadCloser, error) {
-	if ls.Options.LuaFileSystem != nil {
-		return ls.Options.LuaFileSystem.Open(path)
-	}
-	return os.Open(path)
-}
-
-func (ls *LState) Stat(luapath string) (os.FileInfo, error) {
-	if ls.Options.LuaFileSystem != nil {
-		return ls.Options.LuaFileSystem.Stat(luapath)
-	}
-	return os.Stat(luapath)
-}
-
-/* }}} */
-
 /* Global {{{ */
 
 func newGlobal() *Global {
@@ -711,6 +690,13 @@ func newLState(options Options) *LState {
 	ls.reg = newRegistry(ls, options.RegistrySize, options.RegistryGrowStep, options.RegistryMaxSize, al)
 	ls.Env = ls.G.Global
 	return ls
+}
+
+func (ls *LState) openFile(filename string) (fs.File, error) {
+	if ls.Options.LuaFileSystem == nil {
+		return os.Open(filename)
+	}
+	return ls.Options.LuaFileSystem.Open(filename)
 }
 
 func (ls *LState) printReg() {
